services/neuro-snn-engine/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
brian2==2.5.4
norse==1.0.0
kafka-python==2.0.2
psycopg2-binary==2.9.7
pydantic==2.3.0
